Fix size_t/int related warnings

+++ src/scxml/ScXMLTransitionElt.cpp
@@ -324,7 +324,7 @@ ScXMLTransitionElt::setEventAttribute(const char * eventstr)
     if ((ptr != NULL) && (strlen(ptr) == 2)) {
       this->needprefixmatching = TRUE;
       // we'll chop off the pattern matching key and use the boolean
-      const int len = strlen(eventstr) - 1;
+      const size_t len = strlen(eventstr) - 1;
       this->event = new char [ len ];
       strncpy(this->event, eventstr, len - 1);
       this->event[ len - 1 ] = '\0';
@@ -500,8 +500,8 @@ ScXMLTransitionElt::isEventMatch(const ScXMLEvent * eventobj) const
 
   if (this->eventkey == eventid) return TRUE;
 
-  const int keylen = strlen(this->eventkey.getString());
-  if (keylen < static_cast<int>(strlen(eventid.getString()))) {
+  const size_t keylen = strlen(this->eventkey.getString());
+  if (keylen < strlen(eventid.getString())) {
     if ((eventid.getString()[keylen] == '.') &&
         strncmp(this->eventkey.getString(), eventid.getString(), keylen) == 0) {
       return TRUE; // this->event matches up to eventobj token separator
